feat(web): pluralizacao real no resto das telas + theme-color + anti-flash do wizard
Continuacao do pente-fino UX (itens 1, 2 e 3 dos pendentes), a pedido do Felipe.

- Pluralizacao (item 5, sweep): troca o "(s)"/"(ns)"/"(oes)" de template por flexao
  real em ~14 telas. Lado C# usa TextHelpers.Plural/PluralWord (param alargado p/
  long, aceita int e long dos contadores); lado JS ganha fmt.plural/fmt.pluralWord
  em format.js. Telas: CaixaMobile, ClientesMobile, LotesMobile, PedidosMobile,
  ContasAPagar, ContasAReceber, Lotes (Index/Detail/Imprimir), Financeiro,
  MobileDevices, Produtos/Form (resumo do wizard). Atributos (title/delta-meta) e
  o data-confirm JSON usam variavel Razor p/ evitar aspas aninhadas; alguns ganham
  concordancia verbal ("1 conta em aberto" / "Sera gerada 1 etiqueta").

- theme-color (item 6, complemento): cor da barra do navegador mobile a partir do
  token --navy-700, injetada via JS no _Layout (sem hex hardcoded; respeita o
  RazorViewHygieneTests que bloqueia hex em view).

- Anti-flash do wizard (item 7): x-cloak nas etapas 2/3/4 do Produtos/Form, que
  apareciam juntas no load antes do Alpine inicializar. O scroll-to-top ao avancar
  etapa ja existia (nextStep/prevStep/goToStep).

Build verde + 12/12 arch tests (Category=Architecture).

+++ EasyStock.Web/Helpers/TextHelpers.cs
@@ -7,11 +7,12 @@ namespace EasyStock.Web.Helpers;
 /// </summary>
 public static class TextHelpers
 {
-    /// <summary>Número + palavra flexionada: <c>Plural(1, "item", "itens")</c> => "1 item".</summary>
-    public static string Plural(int n, string singular, string plural)
+    /// <summary>Número + palavra flexionada: <c>Plural(1, "item", "itens")</c> => "1 item".
+    /// Param <c>long</c> aceita int e long por widening (contadores variam).</summary>
+    public static string Plural(long n, string singular, string plural)
         => $"{n} {(n == 1 ? singular : plural)}";
 
     /// <summary>Só a palavra flexionada (sem número), p/ quando o número já é exibido à parte.</summary>
-    public static string PluralWord(int n, string singular, string plural)
+    public static string PluralWord(long n, string singular, string plural)
         => n == 1 ? singular : plural;
 }
